Update pytest markers and remove redundant methods, restructure model dir
