+++ .github/workflows/perf-models-impl.yaml
@@ -66,11 +66,11 @@ jobs:
               - models/demos/yolov8s_world/tests/
               - models/demos/yolov8s/tests
               - models/experimental/yolov6l/tests/
-              - models/demos/mobilenetv2/tests/
+              - models/demos/mobilenetv2/tests/perf/
               - models/demos/vgg_unet/tests/
               - models/demos/yolov9c/tests/perf/
-              - models/demos/yolov10x/tests/
-              - models/experimental/vanilla_unet/test/
+              - models/demos/yolov10x/tests/perf/
+              - models/experimental/vanilla_unet/tests/perf/
 
           - name: tt_transformers
             model-list:
