+++ models/demos/mobilenetv2/tests/pcc/test_mobilenetv2.py
@@ -6,9 +6,9 @@
 import pytest
 
 import ttnn
-from models.demos.mobilenetv2.load_model_utils import load_torch_model
+from models.demos.mobilenetv2.common import load_torch_model
 from models.demos.mobilenetv2.reference.mobilenetv2 import Mobilenetv2
-from models.demos.mobilenetv2.tests.mobilenetv2_common import MOBILENETV2_BATCH_SIZE, MOBILENETV2_L1_SMALL_SIZE
+from models.demos.mobilenetv2.tests.perf.mobilenetv2_common import MOBILENETV2_BATCH_SIZE, MOBILENETV2_L1_SMALL_SIZE
 from models.demos.mobilenetv2.tt import ttnn_mobilenetv2
 from models.demos.mobilenetv2.tt.model_preprocessing import (
     create_mobilenetv2_input_tensors,
