+++ models/demos/mobilenetv2/tests/perf/mobilenetv2_common.py
