+++ models/demos/vgg_unet/README.md
@@ -43,7 +43,7 @@ pytest models/demos/vgg_unet/tests/pcc/test_vgg_unet.py::test_vgg_unet[0-pretrai
 
 Use the following command to run the e2e perf with trace 2cq:
 ```sh
-pytest models/demos/vgg_unet/tests/test_e2e_performant.py
+pytest models/demos/vgg_unet/tests/perf/test_e2e_performant.py
 ```
 - end-2-end perf with Trace+2CQs is 80 FPS
