+++ models/demos/yolov10x/runner/performant_runner_infra.py
@@ -9,7 +9,7 @@ import torch
 from loguru import logger
 
 import ttnn
-from models.demos.yolov10x.load_model_utils import load_torch_model
+from models.demos.yolov10x.common import load_torch_model
 from models.demos.yolov10x.tt.model_preprocessing import create_yolov10x_model_parameters
 from models.demos.yolov10x.tt.yolov10x import TtnnYolov10
 from models.utility_functions import divup, is_wormhole_b0
