+++ models/experimental/vanilla_unet/runner/performant_runner_infra.py
@@ -12,7 +12,7 @@ from models.experimental.vanilla_unet.ttnn.ttnn_unet import TtUnet
 from models.utility_functions import divup, is_wormhole_b0
 from models.experimental.vanilla_unet.test.pcc.test_ttnn_unet import create_custom_preprocessor
 from tests.ttnn.utils_for_testing import assert_with_pcc
-from models.experimental.vanilla_unet.load_model_utils import load_torch_model
+from models.experimental.vanilla_unet.common import load_torch_model
 
 
 class VanillaUNetPerformanceRunnerInfra:
